Refactor (onnx/finn): factor out transforms
